Fix recreation of unique Index
This commit fixes the issue where the unique index is created twice when the field is a one-to-one foreign key.
This was due to the field being marked as 'UNIQUE' and also having a Unique Index created if the foreign key is declared custom through the 'Relation' annotation.
Now, the CLI checks if the foreign key has a unique index created and skips being marked 'UNIQUE' at the column definition.